requirements.txt
# Core Flask dependencies
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Flask-SocketIO==5.3.6
Flask-Babel==4.0.0

# Database
psycopg2-binary==2.9.7
SQLAlchemy==2.0.21

# Redis and Caching
redis>=4.5.2,<5.0.0
Flask-Caching==2.1.0
Flask-Limiter==3.5.0

# Background Tasks
celery==5.3.4
celery[redis]==5.3.4

# HTTP and API
requests==2.31.0
httpx>=0.25.0

# Authentication and Security
bcrypt==4.0.1
cryptography==41.0.7
python-jose[cryptography]==3.3.0

# Configuration and Environment
python-dotenv==1.0.0
pydantic==2.4.2
pydantic-settings==2.0.3

# OpenAI and AI
openai==1.3.5
tiktoken==0.5.1

# Stripe Integration
stripe==7.7.0

# Google APIs
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0

# Document Processing
PyPDF2==3.0.1
python-docx==0.8.11
beautifulsoup4==4.12.2
lxml==4.9.3

# Vector Database and Embeddings
numpy==1.25.2
scikit-learn==1.3.2

# Telegram Bot
python-telegram-bot==20.7

# Utilities
python-dateutil==2.8.2
pytz==2023.3
marshmallow==3.20.1

# Development and Testing
pytest==7.4.3
pytest-flask==1.3.0
pytest-cov==4.1.0
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# Logging and Monitoring
structlog==23.2.0
prometheus-client==0.19.0
psutil==5.9.6

# Production WSGI Server
gunicorn==21.2.0
PyYAML==6.0.1